Change libflashsupport install,	add oss_hdaudio option and misc.
A) Change libflashsupport install method to try and support both 32 and 64 bit copies of the library. Not sure this is the best fix, but this will have to be dealt with eventually given the Flash V10 64-bit port.
B) Add an hdaudio_noskip option to disable some checks for unconnected jacks (I know it helps Ionic).
C) Add an empty soundon.user file under Linux/FreeBSD. By default, it's disabled since it doesn't have executable permissions.
D) Undo rev. 502 change to oss_envy24 since it seems to cause a crash[1]:
E) Other misc. changes.

[1] http://4front-tech.com/forum/viewtopic.php?p=11212#11212

+++ cmd/ossplay/ossplay_console.c
@@ -101,7 +101,7 @@ print_record_update (int v, double secs, const char * fname, int update)
 
       hours = mins / 60;
       mins = mins % 60;
-      fprintf (stderr, "] %02d:%02d:%02d", hours, mins, secs);
+      fprintf (stderr, "] %02d:%02d:%02d", hours, mins, (int)secs);
     }
 
   if (!level_meters)
